Switch new and old status pages
