+++ site/frontend/src/pages/status/collector.vue
