+++ site/frontend/src/urls.ts
@@ -4,7 +4,7 @@ export const INFO_URL = `${BASE_URL}/info`;
 
 export const DASHBOARD_DATA_URL = `${BASE_URL}/dashboard`;
 export const STATUS_DATA_URL = `${BASE_URL}/status_page`;
-export const STATUS_DATA_NEW_URL = `${BASE_URL}/status_page_new`;
+export const STATUS_DATA_OLD_URL = `${BASE_URL}/status_page_old`;
 export const BOOTSTRAP_DATA_URL = `${BASE_URL}/bootstrap`;
 export const GRAPH_DATA_URL = `${BASE_URL}/graphs`;
 export const COMPARE_DATA_URL = `${BASE_URL}/get`;
