+++ site/frontend/templates/layout.html
@@ -10,14 +10,14 @@
   {% block head %}{% endblock %}
 </head>
 <body class="container">
-  <div>&gt; <a href="index.html">graphs</a>, <a href="compare.html">compare</a>,
-    <a href="dashboard.html">dashboard</a>, <a href="bootstrap.html">bootstrap</a>,
-    <a href="status.html">status</a>, <a href="status_new.html">status new</a>, <a href="help.html">help</a>.
-  </div>
-  {% block content %}{% endblock %}
-  <div style="text-align: center;">
-    <a href="https://github.com/rust-lang/rustc-perf">Contribute on GitHub</a>
-  </div>
-  {% block script %}{% endblock %}
+<div>&gt; <a href="index.html">graphs</a>, <a href="compare.html">compare</a>,
+  <a href="dashboard.html">dashboard</a>, <a href="bootstrap.html">bootstrap</a>,
+  <a href="status.html">status</a>, <a href="status_old.html">status (old)</a>, <a href="help.html">help</a>.
+</div>
+{% block content %}{% endblock %}
+<div style="text-align: center;">
+  <a href="https://github.com/rust-lang/rustc-perf">Contribute on GitHub</a>
+</div>
+{% block script %}{% endblock %}
 </body>
 </html>
