+++ site/frontend/templates/pages/status_old.html
@@ -1,10 +1,10 @@
 {% extends "layout.html" %}
 {% block head %}
-<link rel="stylesheet" type="text/css" href="scripts/status_new.css">
+<link rel="stylesheet" type="text/css" href="scripts/status_old.css">
 {% endblock %}
 {% block content %}
 <div id="app"></div>
 {% endblock %}
 {% block script %}
-<script src="scripts/status_new.js"></script>
+<script src="scripts/status_old.js"></script>
 {% endblock %}
