+++ site/src/request_handlers/status_page_new.rs
@@ -10,7 +10,7 @@ use hashbrown::HashMap;
 use std::sync::Arc;
 use std::time::Duration;
 
-pub async fn handle_status_page_new(ctxt: Arc<SiteCtxt>) -> anyhow::Result<status_new::Response> {
+pub async fn handle_status_page(ctxt: Arc<SiteCtxt>) -> anyhow::Result<status_new::Response> {
     let conn = ctxt.conn().await;
 
     // The queue contains any in-progress request(s) and then the following requests in queue order
